+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added: `partial-no-import` rule.
 - Added: `media-feature-value-dollar-variable` rule.
+- Added: `at-import-partial-extension-blacklist` rule.
 
 # 1.1.1